New option for excluding elements and new ability to read image alt text.
Can now pass in a second selector string for elements to exclude from Read-it-to-Me.  Excluding all tables is a good example of this, as a linear read of tables headings and cells is not a good experience.

Now, if an image has alt text, that text will be read aloud.  This text is bracketed by a small amount of text to help a listener understand what's going on: Image description begin. THEALTTEXT Image description end.